requirements.txt
httpie
markdown
django
django-filter
djangorestframework
social-auth-app-django
